CHK-208 Refactoring finished
