+++ app/src/main/java/com/chaikasoft/app/ui/activities/MainActivity.kt
@@ -19,14 +19,18 @@ import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.semantics.testTagsAsResourceId
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import androidx.navigation.NavHostController
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import com.chaikasoft.app.ui.components.bottombar.BottomBar
+import com.chaikasoft.app.ui.components.bottombar.BottomBarTab
 import com.chaikasoft.app.ui.components.bottombar.HistoricalTripBottomBar
+import com.chaikasoft.app.ui.components.bottombar.HistoricalTripBottomBarTab
 import com.chaikasoft.app.ui.components.bottombar.NavigationBlockedBottomSheet
 import com.chaikasoft.app.ui.components.topbar.MenuItem
 import com.chaikasoft.app.ui.components.topbar.TopBar
 import com.chaikasoft.app.ui.navigation.NavGraph
+import com.chaikasoft.app.ui.navigation.NavigationGuards
 import com.chaikasoft.app.ui.navigation.Routes
 import com.chaikasoft.app.ui.navigation.Screen
 import com.chaikasoft.app.ui.theme.ChaikaTheme
@@ -77,43 +81,63 @@ class MainActivity : ComponentActivity() {
 
                 Scaffold(
                     topBar = {
-                        TopBar(
-                            currentScreen = currentScreen,
-                            currentRoute = currentRoute,
-                            navController = navController,
-                            menuItems = if (currentScreen.showMenuIcon) menuItems else emptyList(),
-                            onBackClick = if (Routes.isHistoricalRoute(currentRoute)) {
-                                { navController.popBackStack(Routes.TRIP_MAIN, inclusive = false) }
-                            } else {
-                                null
-                            },
-                            onMenuItemClick = { item ->
-                                when (item) {
-                                    MenuItem.REFILL -> navController.navigate(
-                                        Routes.PRODUCT_REPLENISH
-                                    )
+                        if (currentRoute !in Routes.routesWithoutTopBar) {
+                            TopBar(
+                                currentScreen = currentScreen,
+                                menuItems = if (currentScreen.showMenuIcon) {
+                                    menuItems
+                                } else {
+                                    emptyList()
+                                },
+                                onBackClick = {
+                                    if (Routes.isHistoricalRoute(currentRoute)) {
+                                        navController.popBackStack(
+                                            Routes.TRIP_MAIN,
+                                            inclusive = false
+                                        )
+                                    } else {
+                                        navController.navigateUp()
+                                    }
+                                },
+                                onMenuItemClick = { item ->
+                                    when (item) {
+                                        MenuItem.REFILL -> navController.navigate(
+                                            Routes.PRODUCT_REPLENISH
+                                        )
 
-                                    MenuItem.AUTONOMOUS_TRIP -> navController.navigate(
-                                        Routes.TRIP_AUTONOMOUS
-                                    )
+                                        MenuItem.AUTONOMOUS_TRIP -> navController.navigate(
+                                            Routes.TRIP_AUTONOMOUS
+                                        )
+                                    }
                                 }
-                            }
-                        )
+                            )
+                        }
                     },
                     bottomBar = {
                         if (Routes.isHistoricalRoute(currentRoute) && historicalShiftUuid != null) {
                             HistoricalTripBottomBar(
-                                navController = navController,
-                                currentRoute = currentRoute,
-                                shiftUuid = historicalShiftUuid
+                                selectedTab = currentHistoricalBottomBarTab(currentRoute),
+                                onTabClick = { tab ->
+                                    navController.navigateToHistoricalTab(
+                                        tab = tab,
+                                        shiftUuid = historicalShiftUuid,
+                                        currentRoute = currentRoute
+                                    )
+                                }
                             )
-                        } else {
+                        } else if (currentRoute != null &&
+                            currentRoute !in Routes.routesWithoutBottomBar
+                        ) {
                             BottomBar(
-                                navController = navController,
-                                currentRoute = currentRoute,
-                                hasActiveShift = hasActiveShift,
-                                onBlockedNavigation = {
-                                    showBlockedNavigationSheet = true
+                                selectedTab = currentBottomBarTab(currentRoute),
+                                onTabClick = { tab ->
+                                    navController.navigateToMainTab(
+                                        tab = tab,
+                                        hasActiveShift = hasActiveShift,
+                                        onBlockedNavigation = {
+                                            showBlockedNavigationSheet = true
+                                        }
+                                    )
                                 }
                             )
                         }
@@ -151,5 +175,72 @@ class MainActivity : ComponentActivity() {
                 )
             }
         }
+    }
+}
+
+private fun currentBottomBarTab(currentRoute: String?): BottomBarTab? = when {
+    Routes.mainRoutes[Routes.TRIP]?.contains(currentRoute) == true -> BottomBarTab.TRIP
+    Routes.mainRoutes[Routes.PRODUCT]?.contains(currentRoute) == true -> BottomBarTab.PRODUCT
+    Routes.mainRoutes[Routes.STATISTICS]?.contains(currentRoute) == true -> BottomBarTab.STATISTICS
+    Routes.mainRoutes[Routes.OPERATION]?.contains(currentRoute) == true -> BottomBarTab.OPERATION
+    Routes.mainRoutes[Routes.PROFILE]?.contains(currentRoute) == true -> BottomBarTab.PROFILE
+    else -> null
+}
+
+private fun currentHistoricalBottomBarTab(currentRoute: String?): HistoricalTripBottomBarTab? =
+    when {
+        Routes.isHistoricalStatisticsRoute(currentRoute) -> HistoricalTripBottomBarTab.STATISTICS
+        Routes.isHistoricalOperationsRoute(currentRoute) -> HistoricalTripBottomBarTab.OPERATIONS
+        else -> null
+    }
+
+private fun NavHostController.navigateToMainTab(
+    tab: BottomBarTab,
+    hasActiveShift: Boolean,
+    onBlockedNavigation: () -> Unit
+) {
+    val route = when (tab) {
+        BottomBarTab.TRIP -> Routes.TRIP_MAIN
+        BottomBarTab.PRODUCT -> Routes.PRODUCT_GRAPH
+        BottomBarTab.STATISTICS -> Routes.STATISTICS_GRAPH
+        BottomBarTab.OPERATION -> Routes.OPERATION_GRAPH
+        BottomBarTab.PROFILE -> Routes.PROFILE
+    }
+    val popUpToRoute = when (tab) {
+        BottomBarTab.TRIP -> Routes.TRIP_GRAPH
+        BottomBarTab.PROFILE -> Routes.PROFILE_GRAPH
+        else -> route
+    }
+
+    if (NavigationGuards.isProtectedBottomGraph(route) && !hasActiveShift) {
+        onBlockedNavigation()
+        return
+    }
+
+    navigate(route) {
+        popUpTo(popUpToRoute) { inclusive = route == popUpToRoute }
+        launchSingleTop = true
+    }
+}
+
+private fun NavHostController.navigateToHistoricalTab(
+    tab: HistoricalTripBottomBarTab,
+    shiftUuid: String,
+    currentRoute: String?
+) {
+    if (currentHistoricalBottomBarTab(currentRoute) == tab) return
+
+    val route = when (tab) {
+        HistoricalTripBottomBarTab.STATISTICS -> Routes.historyStatistics(shiftUuid)
+        HistoricalTripBottomBarTab.OPERATIONS -> Routes.historyOperations(shiftUuid)
+    }
+
+    navigate(route) {
+        popUpTo(Routes.HISTORY_GRAPH) {
+            inclusive = false
+            saveState = true
+        }
+        launchSingleTop = true
+        restoreState = true
     }
 }
