+++ app/src/main/java/com/chaikasoft/app/ui/components/product/CartProductItem.kt
@@ -2,37 +2,17 @@ package com.chaikasoft.app.ui.components.product
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.aspectRatio
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.Close
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.HorizontalDivider
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.constraintlayout.compose.ConstraintLayout
-import androidx.constraintlayout.compose.Dimension
 import com.chaikasoft.app.R
 import com.chaikasoft.app.ui.dto.Product
 import com.chaikasoft.app.ui.theme.ChaikaTheme
@@ -49,86 +29,19 @@ fun CartProductItem(
     onRemoveFromPackage: (() -> Unit)? = null,
     modifier: Modifier = Modifier
 ) {
-    Card(
-        modifier = modifier.fillMaxWidth(),
-        colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surface)
-    ) {
-        ConstraintLayout(
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            val (divTop, divBottom, imageRef, nameRef, priceRowRef) = createRefs()
-
-            // Верхний разделитель
-            HorizontalDivider(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .constrainAs(divTop) {
-                        top.linkTo(parent.top)
-                        start.linkTo(parent.start)
-                        end.linkTo(parent.end)
-                    }
-            )
-
-            // Изображение продукта
-            ProductImage(
-                imageUrl = product.image,
-                modifier = Modifier
-                    .constrainAs(imageRef) {
-                        top.linkTo(divTop.bottom)
-                        bottom.linkTo(parent.bottom)
-                        start.linkTo(parent.start)
-                    }
-                    .size(ProductDimens.CartProductItem.ImageSize)
-                    .padding(end = ProductDimens.PaddingM)
-                    .aspectRatio(1f),
-                contentDescription = product.name
-            )
-
-            // Название продукта
-            Text(
-                text = product.name,
-                style = MaterialTheme.typography.titleMedium,
-                fontWeight = FontWeight.Bold,
-                color = MaterialTheme.colorScheme.onBackground,
-                fontSize = ProductDimens.CartProductItem.NameFontSize,
-                maxLines = ProductDimens.CartProductItem.MaxNameLines,
-                modifier = Modifier.constrainAs(nameRef) {
-                    top.linkTo(imageRef.top)
-                    start.linkTo(imageRef.end)
-                    end.linkTo(parent.end)
-                    width = Dimension.fillToConstraints
-                }
-            )
-
-            // Строка с ценой и действиями
-            CartProductActionsRow(
-                product = product,
-                onQuantityIncrease = onQuantityIncrease,
-                onQuantityDecrease = onQuantityDecrease,
-                onAddToCart = onAddToCart,
-                onRemove = onRemove,
-                onRemoveFromPackage = onRemoveFromPackage,
-                modifier = Modifier.constrainAs(priceRowRef) {
-                    start.linkTo(nameRef.start)
-                    end.linkTo(parent.end)
-                    bottom.linkTo(imageRef.bottom)
-                    width = Dimension.fillToConstraints
-                }
-            )
-
-            // Нижний разделитель
-            HorizontalDivider(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(top = 8.dp)
-                    .constrainAs(divBottom) {
-                        top.linkTo(imageRef.bottom)
-                        start.linkTo(parent.start)
-                        end.linkTo(parent.end)
-                        bottom.linkTo(parent.bottom)
-                    }
-            )
-        }
+    ProductListItemLayout(
+        product = product,
+        modifier = modifier
+    ) { actionRowModifier ->
+        CartProductActionsRow(
+            product = product,
+            onQuantityIncrease = onQuantityIncrease,
+            onQuantityDecrease = onQuantityDecrease,
+            onAddToCart = onAddToCart,
+            onRemove = onRemove,
+            onRemoveFromPackage = onRemoveFromPackage,
+            modifier = actionRowModifier
+        )
     }
 }
 
@@ -147,7 +60,6 @@ private fun CartProductActionsRow(
             .fillMaxWidth()
             .height(ProductDimens.CartProductItem.QuantitySelectorHeight)
     ) {
-        // Цена слева
         Column(
             modifier = Modifier
                 .align(Alignment.CenterStart)
@@ -162,46 +74,28 @@ private fun CartProductActionsRow(
         }
 
         if (product.isInCart) {
-            QuantitySelector(
+            ProductItemQuantitySelector(
                 quantity = product.quantity,
                 onIncrease = onQuantityIncrease,
                 onDecrease = onQuantityDecrease,
                 modifier = Modifier
                     .align(Alignment.CenterEnd)
-                    .size(
-                        width = ProductDimens.CartProductItem.QuantitySelectorWidth,
-                        height = ProductDimens.CartProductItem.QuantitySelectorHeight
-                    ),
-                colorBack = MaterialTheme.colorScheme.surfaceVariant,
-                colorText = MaterialTheme.colorScheme.onSurface
+                    .quantitySelectorSize()
             )
         } else {
-            Button(
+            ProductItemAddButton(
                 modifier = Modifier
                     .align(Alignment.CenterEnd)
-                    .size(
-                        width = ProductDimens.CartProductItem.QuantitySelectorWidth,
-                        height = ProductDimens.CartProductItem.QuantitySelectorHeight
-                    ),
-                shape = CircleShape,
-                contentPadding = PaddingValues(0.dp),
-                colors = ButtonDefaults.buttonColors(
-                    contentColor = MaterialTheme.colorScheme.primary
-                ),
-                onClick = onAddToCart
-            ) {
-                Icon(
-                    imageVector = Icons.Default.Add,
-                    contentDescription = stringResource(id = R.string.cart_product_add_to_cart),
-                    tint = Color.White,
-                    modifier = Modifier.fillMaxSize(0.75F)
-                )
-            }
+                    .quantitySelectorSize(),
+                onClick = onAddToCart,
+                contentDescriptionRes = R.string.cart_product_add_to_cart
+            )
         }
 
         if (product.isInCart) {
-            IconButton(
+            ProductItemRemoveButton(
                 onClick = onRemove,
+                contentDescriptionRes = R.string.cart_product_remove_from_cart,
                 modifier = Modifier
                     .align(Alignment.CenterEnd)
                     .padding(
@@ -209,18 +103,11 @@ private fun CartProductActionsRow(
                         ProductDimens.CartProductItem.QuantitySelectorWidth +
                             ProductDimens.CartProductItem.RemoveButtonPadding
                     )
-            ) {
-                Icon(
-                    imageVector = Icons.Default.Close,
-                    contentDescription = stringResource(
-                        id = R.string.cart_product_remove_from_cart
-                    ),
-                    tint = MaterialTheme.colorScheme.onSurface
-                )
-            }
+            )
         } else if (product.isInCart && onRemoveFromPackage != null) {
-            IconButton(
+            ProductItemRemoveButton(
                 onClick = onRemoveFromPackage,
+                contentDescriptionRes = R.string.cart_product_remove_from_package,
                 modifier = Modifier
                     .align(Alignment.CenterEnd)
                     .offset(
@@ -228,15 +115,7 @@ private fun CartProductActionsRow(
                         (-ProductDimens.CartProductItem.QuantitySelectorWidth) -
                             ProductDimens.CartProductItem.RemoveButtonPadding
                     )
-            ) {
-                Icon(
-                    imageVector = Icons.Default.Close,
-                    contentDescription = stringResource(
-                        id = R.string.cart_product_remove_from_package
-                    ),
-                    tint = MaterialTheme.colorScheme.onSurface
-                )
-            }
+            )
         }
     }
 }
