+++ app/src/main/java/com/chaikasoft/app/ui/components/product/ReplenishProductItem.kt
@@ -2,36 +2,16 @@ package com.chaikasoft.app.ui.components.product
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.aspectRatio
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.Close
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.HorizontalDivider
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.constraintlayout.compose.ConstraintLayout
-import androidx.constraintlayout.compose.Dimension
 import com.chaikasoft.app.R
 import com.chaikasoft.app.ui.dto.Product
 import com.chaikasoft.app.ui.theme.ChaikaTheme
@@ -47,85 +27,19 @@ fun ReplenishProductItem(
     packageQuantity: Int,
     modifier: Modifier = Modifier
 ) {
-    Card(
-        modifier = modifier.fillMaxWidth(),
-        colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surface)
-    ) {
-        ConstraintLayout(
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            val (divTop, divBottom, imageRef, nameRef, priceRowRef) = createRefs()
-
-            // Верхний разделитель
-            HorizontalDivider(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .constrainAs(divTop) {
-                        top.linkTo(parent.top)
-                        start.linkTo(parent.start)
-                        end.linkTo(parent.end)
-                    }
-            )
-
-            // Изображение продукта
-            ProductImage(
-                imageUrl = product.image,
-                modifier = Modifier
-                    .constrainAs(imageRef) {
-                        top.linkTo(divTop.bottom)
-                        bottom.linkTo(parent.bottom)
-                        start.linkTo(parent.start)
-                    }
-                    .size(ProductDimens.CartProductItem.ImageSize)
-                    .padding(end = ProductDimens.PaddingM)
-                    .aspectRatio(1f),
-                contentDescription = product.name
-            )
-
-            // Название продукта
-            Text(
-                text = product.name,
-                style = MaterialTheme.typography.titleMedium,
-                fontWeight = FontWeight.Bold,
-                color = MaterialTheme.colorScheme.onBackground,
-                fontSize = ProductDimens.CartProductItem.NameFontSize,
-                maxLines = ProductDimens.CartProductItem.MaxNameLines,
-                modifier = Modifier.constrainAs(nameRef) {
-                    top.linkTo(imageRef.top)
-                    start.linkTo(imageRef.end)
-                    end.linkTo(parent.end)
-                    width = Dimension.fillToConstraints
-                }
-            )
-
-            StockQuantityRow(
-                product = product,
-                quantityToShow = packageQuantity + product.quantity,
-                onAddToCart = onAddToCart,
-                onQuantityIncrease = onQuantityIncrease,
-                onQuantityDecrease = onQuantityDecrease,
-                onRemove = onRemove,
-                modifier = Modifier.constrainAs(priceRowRef) {
-                    start.linkTo(nameRef.start)
-                    end.linkTo(parent.end)
-                    bottom.linkTo(imageRef.bottom)
-                    width = Dimension.fillToConstraints
-                }
-            )
-
-            // Нижний разделитель
-            HorizontalDivider(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(top = 8.dp)
-                    .constrainAs(divBottom) {
-                        top.linkTo(imageRef.bottom)
-                        start.linkTo(parent.start)
-                        end.linkTo(parent.end)
-                        bottom.linkTo(parent.bottom)
-                    }
-            )
-        }
+    ProductListItemLayout(
+        product = product,
+        modifier = modifier
+    ) { actionRowModifier ->
+        StockQuantityRow(
+            product = product,
+            quantityToShow = packageQuantity + product.quantity,
+            onAddToCart = onAddToCart,
+            onQuantityIncrease = onQuantityIncrease,
+            onQuantityDecrease = onQuantityDecrease,
+            onRemove = onRemove,
+            modifier = actionRowModifier
+        )
     }
 }
 
@@ -144,7 +58,6 @@ private fun StockQuantityRow(
             .fillMaxWidth()
             .height(ProductDimens.CartProductItem.QuantitySelectorHeight)
     ) {
-        // Левый блок: остаток
         Column(
             modifier = Modifier
                 .align(Alignment.CenterStart)
@@ -164,23 +77,18 @@ private fun StockQuantityRow(
         }
 
         if (product.isInCart) {
-            // если товар уже в корзине — показываем селектор количества и кнопку удаления
-            QuantitySelector(
+            ProductItemQuantitySelector(
                 quantity = product.quantity,
                 onIncrease = onQuantityIncrease,
                 onDecrease = onQuantityDecrease,
                 modifier = Modifier
                     .align(Alignment.CenterEnd)
-                    .size(
-                        width = ProductDimens.CartProductItem.QuantitySelectorWidth,
-                        height = ProductDimens.CartProductItem.QuantitySelectorHeight
-                    ),
-                colorBack = MaterialTheme.colorScheme.surfaceVariant,
-                colorText = MaterialTheme.colorScheme.onSurface
+                    .quantitySelectorSize()
             )
 
-            IconButton(
+            ProductItemRemoveButton(
                 onClick = onRemove,
+                contentDescriptionRes = R.string.cart_product_remove_from_cart,
                 modifier = Modifier
                     .align(Alignment.CenterEnd)
                     .padding(
@@ -188,38 +96,16 @@ private fun StockQuantityRow(
                         ProductDimens.CartProductItem.QuantitySelectorWidth +
                             ProductDimens.CartProductItem.RemoveButtonPadding
                     )
-            ) {
-                Icon(
-                    imageVector = Icons.Default.Close,
-                    contentDescription = stringResource(
-                        id = R.string.cart_product_remove_from_cart
-                    ),
-                    tint = MaterialTheme.colorScheme.onSurface
-                )
-            }
+            )
         } else {
-            Button(
+            ProductItemAddButton(
                 modifier = Modifier
                     .align(Alignment.CenterEnd)
-                    .size(
-                        width = ProductDimens.CartProductItem.QuantitySelectorWidth,
-                        height = ProductDimens.CartProductItem.QuantitySelectorHeight
-                    ),
-                shape = CircleShape,
-                contentPadding = PaddingValues(0.dp),
-                colors = ButtonDefaults.buttonColors(
-                    containerColor = MaterialTheme.colorScheme.primary, // яркая красная пилюля
-                    contentColor = Color.White
-                ),
-                onClick = onAddToCart
-            ) {
-                Icon(
-                    imageVector = Icons.Default.Add,
-                    contentDescription = stringResource(id = R.string.cart_product_add_to_cart),
-                    tint = Color.White,
-                    modifier = Modifier.fillMaxSize(0.6F)
-                )
-            }
+                    .quantitySelectorSize(),
+                onClick = onAddToCart,
+                contentDescriptionRes = R.string.cart_product_add_to_cart,
+                iconScale = 0.6F
+            )
         }
     }
 }
