+++ app/src/main/java/com/chaikasoft/app/ui/screens/profile/FeedbackContent.kt
@@ -1,0 +1,242 @@
+package com.chaikasoft.app.ui.screens.profile
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.Button
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.DropdownMenuItem
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.ExposedDropdownMenuBox
+import androidx.compose.material3.ExposedDropdownMenuDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.input.ImeAction
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.unit.dp
+import com.chaikasoft.app.R
+
+@Composable
+internal fun FeedbackContent(
+    isSubmitted: Boolean,
+    categories: List<String>,
+    selectedCategory: String,
+    name: String,
+    email: String,
+    subject: String,
+    message: String,
+    showEmailError: Boolean,
+    isSubmitEnabled: Boolean,
+    onCategoryChange: (String) -> Unit,
+    onNameChange: (String) -> Unit,
+    onEmailChange: (String) -> Unit,
+    onSubjectChange: (String) -> Unit,
+    onMessageChange: (String) -> Unit,
+    onSubmit: () -> Unit,
+    onSendAnother: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    Column(
+        modifier = modifier
+            .fillMaxSize()
+            .padding(16.dp)
+            .verticalScroll(rememberScrollState())
+    ) {
+        if (isSubmitted) {
+            FeedbackSuccessCard(onSendAnother = onSendAnother)
+        } else {
+            FeedbackFormCard(
+                categories = categories,
+                selectedCategory = selectedCategory,
+                name = name,
+                email = email,
+                subject = subject,
+                message = message,
+                showEmailError = showEmailError,
+                isSubmitEnabled = isSubmitEnabled,
+                onCategoryChange = onCategoryChange,
+                onNameChange = onNameChange,
+                onEmailChange = onEmailChange,
+                onSubjectChange = onSubjectChange,
+                onMessageChange = onMessageChange,
+                onSubmit = onSubmit
+            )
+            FeedbackAdditionalInfoCard()
+        }
+    }
+}
+
+@Composable
+private fun FeedbackFormCard(
+    categories: List<String>,
+    selectedCategory: String,
+    name: String,
+    email: String,
+    subject: String,
+    message: String,
+    showEmailError: Boolean,
+    isSubmitEnabled: Boolean,
+    onCategoryChange: (String) -> Unit,
+    onNameChange: (String) -> Unit,
+    onEmailChange: (String) -> Unit,
+    onSubjectChange: (String) -> Unit,
+    onMessageChange: (String) -> Unit,
+    onSubmit: () -> Unit
+) {
+    Card(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(bottom = 16.dp),
+        elevation = CardDefaults.cardElevation(defaultElevation = 4.dp)
+    ) {
+        Column(
+            modifier = Modifier.padding(16.dp)
+        ) {
+            Text(
+                text = stringResource(R.string.feedback_experience_title),
+                style = MaterialTheme.typography.titleMedium,
+                fontWeight = FontWeight.Bold,
+                modifier = Modifier.padding(bottom = 16.dp)
+            )
+            Text(
+                text = stringResource(R.string.feedback_category),
+                style = MaterialTheme.typography.bodyLarge,
+                fontWeight = FontWeight.Medium,
+                modifier = Modifier.padding(bottom = 8.dp)
+            )
+            FeedbackCategoryDropdown(
+                categories = categories,
+                selectedCategory = selectedCategory,
+                onCategoryChange = onCategoryChange
+            )
+
+            Spacer(modifier = Modifier.height(16.dp))
+
+            OutlinedTextField(
+                value = name,
+                onValueChange = onNameChange,
+                label = { Text(stringResource(R.string.feedback_name_label)) },
+                modifier = Modifier.fillMaxWidth(),
+                keyboardOptions = KeyboardOptions(
+                    imeAction = ImeAction.Next
+                )
+            )
+
+            Spacer(modifier = Modifier.height(16.dp))
+
+            OutlinedTextField(
+                value = email,
+                onValueChange = onEmailChange,
+                label = { Text(stringResource(R.string.feedback_email_label)) },
+                modifier = Modifier.fillMaxWidth(),
+                keyboardOptions = KeyboardOptions(
+                    keyboardType = KeyboardType.Email,
+                    imeAction = ImeAction.Next
+                ),
+                isError = showEmailError,
+                supportingText = {
+                    if (showEmailError) {
+                        Text(
+                            text = stringResource(R.string.feedback_email_invalid),
+                            color = MaterialTheme.colorScheme.error
+                        )
+                    }
+                }
+            )
+
+            Spacer(modifier = Modifier.height(16.dp))
+
+            OutlinedTextField(
+                value = subject,
+                onValueChange = onSubjectChange,
+                label = { Text(stringResource(R.string.feedback_subject_label)) },
+                modifier = Modifier.fillMaxWidth(),
+                keyboardOptions = KeyboardOptions(
+                    imeAction = ImeAction.Next
+                )
+            )
+
+            Spacer(modifier = Modifier.height(16.dp))
+
+            OutlinedTextField(
+                value = message,
+                onValueChange = onMessageChange,
+                label = { Text(stringResource(R.string.feedback_message_label)) },
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .height(120.dp),
+                keyboardOptions = KeyboardOptions(
+                    imeAction = ImeAction.Done
+                ),
+                maxLines = 5
+            )
+
+            Spacer(modifier = Modifier.height(24.dp))
+
+            Button(
+                onClick = onSubmit,
+                modifier = Modifier.fillMaxWidth(),
+                enabled = isSubmitEnabled
+            ) {
+                Text(stringResource(R.string.feedback_send_button))
+            }
+        }
+    }
+}
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+private fun FeedbackCategoryDropdown(
+    categories: List<String>,
+    selectedCategory: String,
+    onCategoryChange: (String) -> Unit
+) {
+    var expanded by remember { mutableStateOf(false) }
+
+    ExposedDropdownMenuBox(
+        expanded = expanded,
+        onExpandedChange = { expanded = it }
+    ) {
+        OutlinedTextField(
+            value = selectedCategory,
+            onValueChange = { },
+            readOnly = true,
+            modifier = Modifier
+                .fillMaxWidth()
+                .menuAnchor(),
+            trailingIcon = {
+                ExposedDropdownMenuDefaults.TrailingIcon(expanded = expanded)
+            }
+        )
+        ExposedDropdownMenu(
+            expanded = expanded,
+            onDismissRequest = { expanded = false }
+        ) {
+            categories.forEach { category ->
+                DropdownMenuItem(
+                    text = { Text(category) },
+                    onClick = {
+                        onCategoryChange(category)
+                        expanded = false
+                    }
+                )
+            }
+        }
+    }
+}
